requirements.txt
openai==1.108.1
python-dotenv==1.1.1
python-telegram-bot==21.11.1
